release.nix: zipPath: support ZIPing archive data
zipPath is a function which can take a derivation path and ZIP it,
such that it can be uploaded to release sites. This is used to include
the patched pythondata-cpu-vexriscv from this repository. However,
this doesn't work when the source is already
archived (e.g. tar.gz). Due to a switch from a prebuilt archive to a
Git branch instead, the source attribute now is such an archive (as
returned by fetchFromGitHub) and thus the function needs to understand
how to deal with these inputs as well.

Signed-off-by: Leon Schuermann <leon@is.currently.online>