release.nix: no Nix paths in pythondata-cpu-vexriscv ZIP archive
With the update of nix-litex in 1a5cfbe7e9 ("Update pinned nix-litex
revision"), which has introduced building the VexRiscv CPUs in a Nix
derivation, it is no longer sufficient to simply archive the `src`
derivation of the pythondata-cpu-vexriscv in order to ship a package
containing the patched VexRiscv CPU variants. Because these are now
built through a Nix derivation, the fixupPhase will attempt to patch
interpreter paths in the resulting package. This, however, will not
work when these packages are not instantiated/used as part of a Nix
derivation. Thus, when generating the archive of the patched
pythondata-cpu-vexriscv package, skip the final fixupPhase.

Signed-off-by: Leon Schuermann <leon@is.currently.online>