release.nix: ship patched LiteX package
In addition to the patched pythondata-cpu-vexriscv package, also
provide the corresponding patched litex package as part of the release
artifacts. This makes it possible for non-Nix users to use the patched
package through their regular pip package manager.

To avoid including any Nix-specifics in the generated package, the
build phases are configured to only unpack, patch and repackage it.

Signed-off-by: Leon Schuermann <leon@is.currently.online>