Update LiteX Packages (including VexRiscv PMP fixes)
This is a (long overdue) update of the LiteX packages, which includes
a fix of the VexRiscv PMP NAPOT region handling. This update is
required to support the Tock PMP infrastructure redesign [1].

It further switches to build FPGA bitstreams with Vivado 2022.2, which
is now packaged under a NUR package repository. Furthermore, it
automatically applies the Python overlays for all Python package set
included in the user's Nixpkgs, so new Nixpkgs versions should not be
as likely to break the expressions contained in this repository.

Signed-off-by: Leon Schuermann <leon@is.currently.online>

[1]: https://github.com/tock/tock/pull/3597
[2]: https://github.com/lschuermann/nur-packages